requirements.txt
fastapi ==0.75.0
uvicorn ==0.17.6
sqlalchemy ==1.4.32 
passlib ==1.7.4
bcrypt ==3.2.0
python-jose ==3.3.0
PyMySQL ==1.0.2
python-dotenv ==0.19.2